move prompt-related api calls to new file
